Check if user was returned
Checks if the user was returned from the getUserInfo function before trying to create an account and login.

This allows us to do an extra verification step inside the getUserInfo function in the config file.

An example where this might be useful is when you only want to allow accounts from a specific google organisation to login, this could now be achieved with the following:

```
getUserInfo: async (accessToken: string) => {
        const response = await fetch('https://www.googleapis.com/oauth2/v3/userinfo', {
          headers: { Authorization: `Bearer ${accessToken}` },
        });
        const user = await response.json();

        // Only allow users from specific organisation to login
        if (!user.hd || user.hd !== 'organisation.com') return null;

        return { email: user.email, sub: user.sub };
      },
```